+++ Bodu.Security.Cryptography/test/Security.Cryptography/PaddingStrategyTests.BlockSizeValidation.cs
@@ -7,13 +7,18 @@
 namespace Bodu.Security.Cryptography;
 
 /// <summary>
-/// Probes the <see cref="IPaddingStrategy" /> contract for unexpected exceptions when the supplied
-/// <c>blockSize</c> is invalid. Currently every length-validating <c>Unpad</c> implementation
-/// evaluates <c>input.Length % blockSize</c> before validating the block size — passing
-/// <c>blockSize == 0</c> therefore surfaces a <see cref="DivideByZeroException" /> instead of the
-/// expected <see cref="ArgumentOutOfRangeException" /> documented by every <c>Pad</c> overload.
-/// These tests assert the corrected, contract-aligned behaviour and fail until the bug is fixed.
+/// Pins the <see cref="IPaddingStrategy" /> contract for invalid <c>blockSize</c> values across
+/// both <c>Pad</c> and <c>Unpad</c>. Every implementation must throw
+/// <see cref="ArgumentOutOfRangeException" /> when <c>blockSize &lt;= 0</c> rather than letting the
+/// modulo arithmetic surface a <see cref="DivideByZeroException" /> or silently succeeding for
+/// inputs whose length happens to be a multiple of <c>|blockSize|</c>.
 /// </summary>
+/// <remarks>
+/// Strategies whose <c>Unpad</c> ignores <c>blockSize</c> (<see cref="NoPadding" />,
+/// <see cref="ZeroPadding" />) opt out of the <c>Unpad</c>-side checks via
+/// <see cref="ValidatesBlockSizeOnUnpad" />. Every padding scheme validates <c>blockSize</c> on
+/// <c>Pad</c>, so the <c>Pad</c>-side tests are universally applicable.
+/// </remarks>
 public abstract partial class PaddingStrategyTests<TPadding>
 {
     /// <summary>
@@ -23,6 +28,44 @@ public abstract partial class PaddingStrategyTests<TPadding>
     /// to <see langword="false" /> so the block-size validation tests are skipped for them.
     /// </summary>
     protected virtual bool ValidatesBlockSizeOnUnpad => true;
+
+    /// <summary>
+    /// Verifies that calling <see cref="IPaddingStrategy.Pad" /> with <c>blockSize == 0</c> throws
+    /// <see cref="ArgumentOutOfRangeException" /> rather than <see cref="DivideByZeroException" />
+    /// from the modulo expression. Universally applicable — every padding scheme validates
+    /// <c>blockSize</c> on <c>Pad</c>.
+    /// </summary>
+    [TestMethod]
+    public void Pad_WhenBlockSizeIsZero_ShouldThrowArgumentOutOfRangeException_fix()
+    {
+        var padding = CreatePadding();
+        byte[] input = new byte[BlockSize];
+
+        Assert.ThrowsExactly<ArgumentOutOfRangeException>(() =>
+        {
+            _ = padding.Pad(input, 0);
+        });
+    }
+
+    /// <summary>
+    /// Verifies that calling <see cref="IPaddingStrategy.Pad" /> with a negative <c>blockSize</c>
+    /// throws <see cref="ArgumentOutOfRangeException" /> rather than silently succeeding when the
+    /// input length happens to be a multiple of <c>|blockSize|</c>.
+    /// </summary>
+    [TestMethod]
+    [DataRow(-1)]
+    [DataRow(-16)]
+    [DataRow(int.MinValue)]
+    public void Pad_WhenBlockSizeIsNegative_ShouldThrowArgumentOutOfRangeException_fix(int blockSize)
+    {
+        var padding = CreatePadding();
+        byte[] input = new byte[BlockSize];
+
+        Assert.ThrowsExactly<ArgumentOutOfRangeException>(() =>
+        {
+            _ = padding.Pad(input, blockSize);
+        });
+    }
 
     /// <summary>
     /// Verifies that calling <see cref="IPaddingStrategy.Unpad" /> with <c>blockSize == 0</c>
